+++ confiture-rest-api/src/audits/dto/entities/criterion-result.dto.ts
@@ -1,18 +1,18 @@
 import { ApiProperty } from "@nestjs/swagger";
 import { CriterionResultStatus, CriterionResultUserImpact } from "../../../generated/prisma/enums";
 import { ExampleImageFileDto } from "./example-image-file.dto";
+import { NotCompliantItemDto } from "./not-compliant-item.dto";
 
 export class CriterionResultDto {
   @ApiProperty({ enum: CriterionResultStatus })
   status: CriterionResultStatus;
   compliantComment: string | null;
-  notCompliantComment: string | null;
   @ApiProperty({ enum: CriterionResultUserImpact })
-  userImpact: CriterionResultUserImpact | null;
   exampleImages: ExampleImageFileDto[];
-  quickWin: boolean;
   notApplicableComment: string | null;
   topic: number;
   criterium: number;
   pageId: number;
+
+  notCompliantItems: NotCompliantItemDto[];
 }