+++ confiture-rest-api/src/audits/dto/entities/not-compliant-item.dto.ts
@@ -1,0 +1,11 @@
+import { ApiProperty } from "@nestjs/swagger";
+import { CriterionResultUserImpact } from "src/generated/prisma/enums";
+
+export class NotCompliantItemDto {
+  id: number;
+  title: string | null;
+  comment: string | null;
+  @ApiProperty({ enum: CriterionResultUserImpact })
+  userImpact: CriterionResultUserImpact | null;
+  quickWin: boolean;
+}