+++ confiture-rest-api/src/audits/dto/requests/update-results.dto.ts
@@ -92,33 +92,39 @@ class UpdateResultsItem {
   compliantComment?: string;
 
   /**
-   * @example "Consectetur ad consectetur Lorem id enim sunt amet ea."
-   */
-  @IsString()
-  @IsOptional()
-  notCompliantComment?: string;
-
-  /**
-   * @example "MAJOR"
-   */
-  @IsString()
-  @IsIn(Object.values(CriterionResultUserImpact))
-  @IsOptional()
-  userImpact?: CriterionResultUserImpact;
-
-  /**
-   * Whether the result is easy to fix
-   */
-  @IsBoolean()
-  @IsOptional()
-  quickWin?: boolean;
-
-  /**
    * @example "Officia aliquip aute ipsum in eiusmod ea et."
    */
   @IsString()
   @IsOptional()
   notApplicableComment?: string;
+
+  @IsArray()
+  @ValidateNested({ each: true })
+  @Type(() => CriterionResultNotCompliantItem)
+  notCompliantItems?: CriterionResultNotCompliantItem[];
+}
+
+class CriterionResultNotCompliantItem {
+  // ID
+  @IsNumber()
+  @IsOptional()
+  id?: number;
+
+  @IsString()
+  @IsOptional()
+  title?: string;
+
+  @IsString()
+  @IsOptional()
+  comment?: string;
+
+  @IsIn(Object.values(CriterionResultUserImpact))
+  @IsOptional()
+  userImpact?: CriterionResultUserImpact;
+
+  @IsBoolean()
+  @IsOptional()
+  quickWin?: boolean;
 }
 
 export class UpdateResultsDto {