+++ confiture-rest-api/src/audits/not-compliant-items/dto/create-not-compliant-item.dto.ts
@@ -1,0 +1,20 @@
+import { IsBoolean, IsEnum, IsOptional, IsString } from "class-validator";
+import { CriterionResultUserImpact } from "../../../generated/prisma/client";
+
+export class CreateNotCompliantItemDto {
+  @IsString()
+  @IsOptional()
+  title?: string | null;
+
+  @IsString()
+  @IsOptional()
+  comment?: string | null;
+
+  @IsEnum(Object.values(CriterionResultUserImpact))
+  @IsOptional()
+  userImpact?: CriterionResultUserImpact | null;
+
+  @IsBoolean()
+  @IsOptional()
+  quickWin?: boolean;
+}