+++ confiture-web-app/src/types/types.ts
@@ -127,8 +127,26 @@ export type NotesFile = components["schemas"]["NotesFileDto"];
 /** Image file attached to specific criterium result when not compliant. */
 export type ExampleImageFile = components["schemas"]["ExampleImageFileDto"];
 
+export type NotCompliantItem = {
+  id: number;
+  title?: string;
+  comment: string | null;
+  userImpact: CriterionResultUserImpact | null;
+  quickWin: boolean;
+};
+
+export type CreateNotCompliantItemData = paths["/audits/{uniqueId}/pages/{slug}/results/{topic}.{criterium}/not-compliant-items"]["post"]["requestBody"]["content"]["application/json"];
+
+export type UpdateNotCompliantItemData = paths["/audits/{uniqueId}/pages/{slug}/results/{topic}.{criterium}/not-compliant-items/{itemId}"]["patch"]["requestBody"]["content"]["application/json"];
+
+export type PatchNotCompliantItemData = { id: number } & Partial<
+  Omit<NotCompliantItem, "id">
+>;
+
 export interface CriteriumResult {
   // ID
+  id: number;
+
   topic: number;
   criterium: number;
   pageId: number;
@@ -137,11 +155,9 @@ export interface CriteriumResult {
   status: CriteriumResultStatus;
 
   compliantComment: string | null;
-  notCompliantComment: string | null;
-  userImpact: CriterionResultUserImpact | null;
   notApplicableComment: string | null;
   exampleImages: ExampleImageFile[];
-  quickWin: boolean;
+  notCompliantItems: NotCompliantItem[];
 }
 
 export enum StoreName {